+++ opentelemetry-appender-tracing/CHANGELOG.md
@@ -42,7 +42,7 @@ transparent to most users.
   implementations (SDK, processor, exporters) to leverage this additional
   information to determine if an event is enabled.
 
-- `u64` and `usize` values are stored as `opentelemetry::logs::AnyValue::Int`
+- `u64`, `i128`, `u128` and `usize` values are stored as `opentelemetry::logs::AnyValue::Int`
 when conversion is feasible. Otherwise stored as
 `opentelemetry::logs::AnyValue::String`. This avoids unnecessary string
 allocation when values can be represented in their original types.